ai/requirements.txt
# --- Web Framework & Server ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
starlette==0.41.3
python-multipart==0.0.20

# --- Async HTTP Client ---
httpx==0.27.2

# --- Database & ORM ---
sqlalchemy[asyncio]==2.0.36
asyncmy==0.2.9
PyMySQL==1.1.1

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
cryptography==44.0.0

# --- Environment Configuration ---
python-dotenv==1.0.1

# --- AI/ML Core Libraries ---
# Using CPU-only PyTorch for smaller image size
torch==2.5.1
sentence-transformers==3.3.1
numpy==1.26.4

# --- LangChain & RAG ---
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langchain-core==0.3.28

# --- Vector Database ---
chromadb==0.5.23

# --- OpenAI API ---
openai==1.58.1

# --- HTTP Requests ---
requests==2.32.3

# --- Data Validation ---
pydantic==2.10.4
pydantic-core==2.27.2

# --- Utilities ---
python-dateutil==2.9.0.post0
tqdm==4.67.1

# --- Compatibility Pins ---
# Resolve urllib3 conflict between requests and kubernetes (chromadb dependency)
urllib3>=1.26.0,<2.4.0

# Ensure compatible typing-extensions
typing-extensions>=4.12.0

# Pin certifi for security
certifi>=2024.8.30
